requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
firebase-admin==5.0.3
pydantic==1.8.2
python-dotenv==0.19.0
openai==0.10.2
langchain